requirements.txt
PyTDC
torch
streamlit
pandas
numpy
scikit-learn
